requirements.txt
 Adafruit_IO==2.5.0
passlib==1.7.4
pymongo==3.11.4
adafruit_bme280==1.0.1
adafruit_pca9685==1.0.1
board==1.0
paho_mqtt==1.5.1
requests==2.25.1
WTForms==2.3.3
Flask_SocketIO==5.1.0
Flask_SQLAlchemy==2.5.1
Flask==1.1.2
Flask_WTF==0.15.1
login==0.0.6
flask_login==0.5.0
paho_mqtt==1.5.1
PyYAML==5.4.1
gunicorn==19.9.0
dnspython==1.16.0
logger==1.4
eventlet==0.30.2
